+++ tests/test_edgetts_voices.py
@@ -2,9 +2,13 @@ import sys
 import os
 
 addon_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.append(addon_dir)
+sys.path.insert(0, addon_dir)
+sys.path.insert(0, os.path.join(addon_dir, 'external'))
 
-from tests.conftest import MockAnkiUtils
+from tests import mock_anki
+
+mock_anki.mock_all()
+
 import superfreetss_addon.servicemanager as servicemanager
 
 def main():